@@ -4,18 +4,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.16.2",
-    "create-react-class": "^15.6.2",
+    "axios": "^0.19.0",
     "font-awesome": "^4.7.0",
     "history": "^4.7.2",
     "immutability-helper": "^2.4.0",
-    "moment": "^2.19.2",
+    "moment": "^2.24.0",
     "prop-types": "^15.6.0",
-    "react": "^16.0.0",
-    "react-barcode": "^1.3.2",
+    "react": "^16.12.0",
     "react-bootstrap": "^0.31.5",
-    "react-burger-menu": "^2.1.9",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.12.0",
     "react-fontawesome": "^1.6.1",
     "react-redux": "^5.0.6",
     "react-router": "^4.2.0",
@@ -33,8 +30,6 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "build": "react-scripts build"
   }
 }
