@@ -1,5 +1,5 @@
 import request from '../utils/request';
-import { REMOTE_URL } from '../config/app';
+import { REMOTE_URL } from '../config';
 
 export function authenticate(){
   return function(dispatch, getState){
