@@ -1,70 +1,55 @@
-import React, { Component } from 'react';
+import React from 'react'
 
-class License extends Component {
-
-  render() {
-    return (
-      <div id="license" className="section">
-
-          <div className="container">
-
-            <div className="block-title-holder">
-              <div className="col-md-12">
-                <h1 className="block-title">Vásárlási feltételek</h1>
-              </div>
-            </div>
-
-
-            <div className="block-content">
-              <div className="col-md-12">
-                <div className="wyswig">
-                  <p>Az Vásárlási Feltételek tartalmazzák Pellicano Marco Allesandro (2000, Szentendre Paprikabíró utca 2, 67352816 - 1 - 33), mint szolgáltató (“Szolgáltató”) által üzemeltetett webáruház használatára vonatkozó általános szerződési feltételeket. Kérjük, hogy csak akkor vegye igénybe szolgáltatásainkat, amennyiben minden pontjával egyetért, és kötelező érvényűnek tekinti magára nézve. Jelen dokumentum nem kerül iktatásra, kizárólag elektronikus formában kerül megkötésre (nem minősül írásba foglalt szerződésnek), magatartási kódexre nem utal. Az dokumentum letölthető az alábbi linkről: <a href="https://jegyek.mobiljeg.hu/vasarlasi-feltetelek">https://jegyek.mobiljeg.hu/vasarlasi-feltetelek</a></p>
-                  <h2>Szolgáltató adatai</h2>
-                  <ul>
-                    <li><span>A szolgáltató neve:</span> Pellicano Marco Allesandro</li> 
-                    <li><span>A szolgáltató székhelye:</span> 2000, Szentendre Paprikabíró utca 2</li> 
-                    <li><span>Elektronikus levelezési címe:</span> mobiljeg@gmail.com</li> 
-                    <li><span>Adószáma:</span> 67352816 - 1 - 33</li> 
-                    <li><span>Telefonszám:</span> +36 30 399 0414</li> 
-                    <li><span>Adatvédelmi nyilvántartási száma:</span> </li> 
-                    <li><span>A szerződés nyelve:</span> magyar</li> 
-                  </ul>
-                  <h2>Alapvető rendelkezések</h2>
-                  <p>A jelen Szabályzatban nem szabályozott kérdésekre, valamint jelen Szabályzat értelmezésére a magyar jog az irányadó, különös tekintettel a Polgári Törvénykönyvről szóló 2013. évi V. törvény („Ptk.”) és az elektronikus kereskedelmi szolgáltatások, valamint az információs társadalommal összefüggő szolgáltatások egyes kérdéseiről szóló 2001. évi CVIII. törvény vonatkozó rendelkezéseire. A vonatkozó jogszabályok kötelező rendelkezései a felekre külön kikötés nélkül is irányadók.</p>
-                  <h2>Jegy vásálás folyamata</h2>
-                  <p>A vásárlónak lehetősége van belépő jegyet vásárolni az "Szolgáltató" jégpályáira online. Vásárlási szándéka esetén az űrlapot kitöltve, a jegy darabszámát kiválasztva elindíthatja a vásárlási folyamatot.</p>
-                  <p>A vásárlás gombra kattintva a folyamat megkezdődik és a vásálót átirányítjuk az online fizetési rendszerbe.</p>
-                  <p>Sikeres tranzakció esetén a vásárló vissztér a webárúházba ahol láthatja a megvásárolt jegy és a tranzakció adatait.</p>
-                  <p>Sikertelen vásárlás esetén a vásárló visszatér a webárúházba ahol látja a meghiúsult tranzakció részleteit.</p>
-                  <h2>Áraink</h2>
-                  <p>Amennyiben a webáruházban hiba vagy hiányosság lép fel fenntartjuk a jogot a korrekcióra. Ilyen esetben a hiba felismerése illetve módosítása után azonnal tájékoztatjuk a vevőt az új adatokról. A vevő ezt követően még egyszer megerősítheti a megrendelést, vagy lehetőség van arra, hogy bármely fél elálljon a szerződéstől.</p>
-                  <h2>A rendelések feldolgozása</h2>
-                  <p>A rendelések feldolgozása azonnal megtörténik fizetés után egyből érvényesítheti a kapott jegyet.</p>
-                  <h2>Adatbeviteli hibák javításának lehetősége</h2>
-                  <p>Az adatbeviteli hibák javítására a „Megrendelés elküldése” gomb megnyomása előtt van lehetősége.</p>
-                  <h2>Visszaigazolás</h2>
-                  <p>Minden rendelésről e-mailes visszajelzést küldünk. Ez azt jelenti, hogy a sikeres vásárlást követően egy automatikus e-mailt kap a rendelés beérkezéséről ami tartalmazza a tranzakció és a megvásárolt jegy adatait.</p>
-                  <h2>Fizetési feltételek</h2>
-                  <p>A "Szolgáltató" webáruházban bankkártyás fizetésre van lehetőség.</p>
-                  <h2>Szerződéskötés feltételeire vonatkozó tájékoztatás</h2>
-                  <p>A szerződéskötés nyelve magyar, a megkötött szerződés írásban megkötöttnek minősül.</p>
-                  <h2>Elállási jog / Visszaszolgáltatási garancia</h2>
-                  <p>A megrendelő a szerződéstől 15 napon belül indokolás nélkül elállhat. A megrendelő az elállás jogát attól a naptól kezdve gyakorolhatja, amikor a jegyet megvette. A "Szolgáltató" köteles a megrendelő által kifizetett összeget haladéktalanul, de legkésőbb az elállást követő harminc napon belül visszatéríteni.</p>
-                  <p>Az elállási jog gyakorlására vonatkozó 45/2014. (II.26) Kormányrendelet szövegét letöltheti a Nemzeti Fogyasztóvédelmi Hatóság honlapjáról.</p>
-                  <h2>Megrendelés módosítása, törlése</h2>
-                  <p>A 2001. évi CVIII. törvény értelmében a megrendelés leadásakor a vásárló felé azonnal értesítést küld az eladó webáruház szoftvere a rendelés felvételének tényéről. Ez az értesítés nem minősül az eladó és a vásárló között létrejött szerződésnek! Csupán jelzi a vásárlónak, hogy rendelési igényét rendszerünk regisztrálta és továbbította az eladó illetékes munkatársa felé.</p>
-                  <p>Amennyiben a vásárlóhoz ez 48 órán belül nem érkezik meg, akkor a vásárló ajánlattételi kötöttségei megszűnnek.</p>
-                  <p>Az eladó a rendelés teljesítésének megkezdéséig megadja a lehetőséget a vásárlónak rendelése visszavonására emailban. A rendelés teljesítésének megkezdésekor a vásárló e-mailben és/vagy telefonon értesítést kap a teljesítés várható időtartamáról, ill. a teljesítés megkezdésének tényéről. Ezután már csak személyesen vagy a “Kapcsolat” menüpontok alatt található elérhetőségek valamelyikén keresztül van lehetőség rendelésének visszavonására.</p>
-                  <p>A megrendelés utólagos módosítására csak írásos formában, e-mailben van lehetőség.</p>
-                </div>
-              </div>
-            </div>
-
+export default function License() {
+  return (
+    <div id="license" className="section">
+      <div className="container">
+        <div className="block-title-holder">
+          <h1 className="block-title">Vásárlási feltételek</h1>
         </div>
-
+        <div className="block-content">
+          <div className="wyswig">
+            <p>Az Vásárlási Feltételek tartalmazzák Pellicano Marco Allesandro (2000, Szentendre Paprikabíró utca 2, 67352816 - 1 - 33), mint szolgáltató (“Szolgáltató”) által üzemeltetett webáruház használatára vonatkozó általános szerződési feltételeket. Kérjük, hogy csak akkor vegye igénybe szolgáltatásainkat, amennyiben minden pontjával egyetért, és kötelező érvényűnek tekinti magára nézve. Jelen dokumentum nem kerül iktatásra, kizárólag elektronikus formában kerül megkötésre (nem minősül írásba foglalt szerződésnek), magatartási kódexre nem utal. Az dokumentum letölthető az alábbi linkről: <a href="https://mobiljeg.hu/jegyek/vasarlasi-feltetelek">https://mobiljeg.hu/jegyek/vasarlasi-feltetelek</a></p>
+            <h2>Szolgáltató adatai</h2>
+            <ul>
+              <li><span>A szolgáltató neve:</span> Pellicano Marco Allesandro</li> 
+              <li><span>A szolgáltató székhelye:</span> 2000, Szentendre Paprikabíró utca 2</li> 
+              <li><span>Elektronikus levelezési címe:</span> mobiljeg@gmail.com</li> 
+              <li><span>Adószáma:</span> 67352816 - 1 - 33</li> 
+              <li><span>Telefonszám:</span> +36 30 399 0414</li> 
+              <li><span>Adatvédelmi nyilvántartási száma:</span> </li> 
+              <li><span>A szerződés nyelve:</span> magyar</li> 
+            </ul>
+            <h2>Alapvető rendelkezések</h2>
+            <p>A jelen Szabályzatban nem szabályozott kérdésekre, valamint jelen Szabályzat értelmezésére a magyar jog az irányadó, különös tekintettel a Polgári Törvénykönyvről szóló 2013. évi V. törvény („Ptk.”) és az elektronikus kereskedelmi szolgáltatások, valamint az információs társadalommal összefüggő szolgáltatások egyes kérdéseiről szóló 2001. évi CVIII. törvény vonatkozó rendelkezéseire. A vonatkozó jogszabályok kötelező rendelkezései a felekre külön kikötés nélkül is irányadók.</p>
+            <h2>Jegy vásálás folyamata</h2>
+            <p>A vásárlónak lehetősége van belépő jegyet vásárolni az "Szolgáltató" jégpályáira online. Vásárlási szándéka esetén az űrlapot kitöltve, a jegy darabszámát kiválasztva elindíthatja a vásárlási folyamatot.</p>
+            <p>A vásárlás gombra kattintva a folyamat megkezdődik és a vásálót átirányítjuk az online fizetési rendszerbe.</p>
+            <p>Sikeres tranzakció esetén a vásárló vissztér a webárúházba ahol láthatja a megvásárolt jegy és a tranzakció adatait.</p>
+            <p>Sikertelen vásárlás esetén a vásárló visszatér a webárúházba ahol látja a meghiúsult tranzakció részleteit.</p>
+            <h2>Áraink</h2>
+            <p>Amennyiben a webáruházban hiba vagy hiányosság lép fel fenntartjuk a jogot a korrekcióra. Ilyen esetben a hiba felismerése illetve módosítása után azonnal tájékoztatjuk a vevőt az új adatokról. A vevő ezt követően még egyszer megerősítheti a megrendelést, vagy lehetőség van arra, hogy bármely fél elálljon a szerződéstől.</p>
+            <h2>A rendelések feldolgozása</h2>
+            <p>A rendelések feldolgozása azonnal megtörténik fizetés után egyből érvényesítheti a kapott jegyet.</p>
+            <h2>Adatbeviteli hibák javításának lehetősége</h2>
+            <p>Az adatbeviteli hibák javítására a „Megrendelés elküldése” gomb megnyomása előtt van lehetősége.</p>
+            <h2>Visszaigazolás</h2>
+            <p>Minden rendelésről e-mailes visszajelzést küldünk. Ez azt jelenti, hogy a sikeres vásárlást követően egy automatikus e-mailt kap a rendelés beérkezéséről ami tartalmazza a tranzakció és a megvásárolt jegy adatait.</p>
+            <h2>Fizetési feltételek</h2>
+            <p>A "Szolgáltató" webáruházban bankkártyás fizetésre van lehetőség.</p>
+            <h2>Szerződéskötés feltételeire vonatkozó tájékoztatás</h2>
+            <p>A szerződéskötés nyelve magyar, a megkötött szerződés írásban megkötöttnek minősül.</p>
+            <h2>Elállási jog / Visszaszolgáltatási garancia</h2>
+            <p>A megrendelő a szerződéstől 15 napon belül indokolás nélkül elállhat. A megrendelő az elállás jogát attól a naptól kezdve gyakorolhatja, amikor a jegyet megvette. A "Szolgáltató" köteles a megrendelő által kifizetett összeget haladéktalanul, de legkésőbb az elállást követő harminc napon belül visszatéríteni.</p>
+            <p>Az elállási jog gyakorlására vonatkozó 45/2014. (II.26) Kormányrendelet szövegét letöltheti a Nemzeti Fogyasztóvédelmi Hatóság honlapjáról.</p>
+            <h2>Megrendelés módosítása, törlése</h2>
+            <p>A 2001. évi CVIII. törvény értelmében a megrendelés leadásakor a vásárló felé azonnal értesítést küld az eladó webáruház szoftvere a rendelés felvételének tényéről. Ez az értesítés nem minősül az eladó és a vásárló között létrejött szerződésnek! Csupán jelzi a vásárlónak, hogy rendelési igényét rendszerünk regisztrálta és továbbította az eladó illetékes munkatársa felé.</p>
+            <p>Amennyiben a vásárlóhoz ez 48 órán belül nem érkezik meg, akkor a vásárló ajánlattételi kötöttségei megszűnnek.</p>
+            <p>Az eladó a rendelés teljesítésének megkezdéséig megadja a lehetőséget a vásárlónak rendelése visszavonására emailban. A rendelés teljesítésének megkezdésekor a vásárló e-mailben és/vagy telefonon értesítést kap a teljesítés várható időtartamáról, ill. a teljesítés megkezdésének tényéről. Ezután már csak személyesen vagy a “Kapcsolat” menüpontok alatt található elérhetőségek valamelyikén keresztül van lehetőség rendelésének visszavonására.</p>
+            <p>A megrendelés utólagos módosítására csak írásos formában, e-mailben van lehetőség.</p>
+          </div>
+        </div>
       </div>
-    );
-  }
+    </div>
+  )
 }
-
-export default License;
