@@ -1,28 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react'
 
-import LoginForm from '../../components/LoginForm/LoginForm';
-import RegistrationForm from '../../components/RegistrationForm/RegistrationForm';
+import { LoginForm } from '../../views'
 
-import './Login.css';
+import './Login.css'
 
-class Login extends Component {
-
-  render() {
-    return (
-      <div id="login" className="section">
-
-        <div className="container">
-
-          <div className="login-forms">
-            <LoginForm />
-            <RegistrationForm />
-          </div>
-
-        </div>
-
+export default function Login() {
+  return (
+    <div id="login" className="section">
+      <div className="container">
+        <LoginForm />
       </div>
-    );
-  }
+    </div>
+  )
 }
-
-export default Login
