@@ -1,36 +1,22 @@
-import React, { Component } from 'react';
+import React from 'react'
 
-class Privacy extends Component {
-
-  render() {
-    return (
-      <div id="privacy" className="section">
-
-          <div className="container">
-
-            <div className="block-title-holder">
-              <div className="col-md-12">
-                <h1 className="block-title">Adatvédelmi nyilatkozat</h1>
-              </div>
+export default function Privacy() {
+  return (
+    <div id="privacy" className="section">
+        <div className="container">
+          <div className="block-title-holder">
+            <h1 className="block-title">Adatvédelmi nyilatkozat</h1>
+          </div>
+          <div className="block-content">
+            <div className="wyswig">
+              <p>Az Ön személyes adatainak védelme a Mobiljég számára kiemelt fontosságú. Weboldalunk használatakor szükséges, azonosításra alkalmas, személyes jellegű adatok begyűjtése és feldolgozása megfelel a jelenleg érvényben lévő magyar előírásoknak ( 2011. évi CXII. törvény).</p>
+              <p>Az Ön adatait bizalmasan kezeljük és harmadik fél számára nem továbbítjuk, kivéve azt az esetet, amikor ez a szerződés teljesítéséhez elengedhetetlen (pl. online fizetés). A harmadik fél a Mobiljég-től kapott adatokat semmilyen formában nem jogosult felhasználni, tárolni, illetve más részére továbbadni. Adatai kezelésekor a Személyes adatok védelméről és a közérdekű adatok nyilvánosságáról szóló 1992. évi LXIII törvénynek megfelelően járunk el.</p>
+              <h2>Az információk tulajdonjoga</h2>
+              <p>Kijelentjük, hogy a Mobiljég birtokolja az oldalainkon látható információkat. A https://mobiljeg.hu oldalainak tartalmát és a kialakítását nemzetközi és magyar törvények védik. Tilos a tőlünk származó információk, cikkek árusítása, bármilyen formában történő megváltoztatása és ismételt megjelentetése. A megjelent anyaggal kapcsolatos minden további jogunkat fenntartjuk, és szükség esetén akár bírósági úton is érvényesítjük.</p>
+              <p>Aki a webáruházunkba regisztrálja magát, az adatvédelmi nyilatkozatunkat elfogadottnak tekinti, azzal egyetért.</p>
             </div>
-
-            <div className="block-content">
-              <div className="col-md-12">
-                <div className="wyswig">
-                  <p>Az Ön személyes adatainak védelme a Mobiljég számára kiemelt fontosságú. Weboldalunk használatakor szükséges, azonosításra alkalmas, személyes jellegű adatok begyűjtése és feldolgozása megfelel a jelenleg érvényben lévő magyar előírásoknak ( 2011. évi CXII. törvény).</p>
-                  <p>Az Ön adatait bizalmasan kezeljük és harmadik fél számára nem továbbítjuk, kivéve azt az esetet, amikor ez a szerződés teljesítéséhez elengedhetetlen (pl. online fizetés). A harmadik fél a Mobiljég-től kapott adatokat semmilyen formában nem jogosult felhasználni, tárolni, illetve más részére továbbadni. Adatai kezelésekor a Személyes adatok védelméről és a közérdekű adatok nyilvánosságáról szóló 1992. évi LXIII törvénynek megfelelően járunk el.</p>
-                  <h2>Az információk tulajdonjoga</h2>
-                  <p>Kijelentjük, hogy a Mobiljég birtokolja az oldalainkon látható információkat. A https://mobiljeg.hu oldalainak tartalmát és a kialakítását nemzetközi és magyar törvények védik. Tilos a tőlünk származó információk, cikkek árusítása, bármilyen formában történő megváltoztatása és ismételt megjelentetése. A megjelent anyaggal kapcsolatos minden további jogunkat fenntartjuk, és szükség esetén akár bírósági úton is érvényesítjük.</p>
-                  <p>Aki a webáruházunkba regisztrálja magát, az adatvédelmi nyilatkozatunkat elfogadottnak tekinti, azzal egyetért.</p>
-                </div>
-              </div>
-            </div>
-
-        </div>
-
+          </div>
       </div>
-    );
-  }
+    </div>
+  )
 }
-
-export default Privacy;
