@@ -1,27 +1,18 @@
-import React, { Component } from 'react';
-import { connect } from "react-redux";
+import React from 'react'
+import { connect } from "react-redux"
 
-class Profil extends Component {
-
-  render() {
-    return (
-      <div id="profil" className="section">
-
-          <div className="container">
-
-            <div className="block-title-holder">
-              <div className="col-md-12">
-                <h1 className="block-title">Profil</h1>
-              </div>
-            </div>
-
+function Profil() {
+  return (
+    <div id="profil" className="section">
+      <div className="container">
+        <div className="block-title-holder">
+          <h1 className="block-title">Profil</h1>
         </div>
-         
       </div>
-    );
-  }
+    </div>
+  )
 }
 
 export default connect(state => ({
   user: state.user
-}))(Profil);
+}))(Profil)
