@@ -1,6 +1,6 @@
-import createHistory from 'history/createBrowserHistory';
+import createHistory from 'history/createBrowserHistory'
 
 // Create a history of your choosing (we're using a browser history in this case)
 const history = createHistory({ basename: '/jegyek' })
 
-export default history;
+export default history
